Asession cart ID cookie
We have our `AddToCart` component and the begining of the `addItemToCart` action. What we want to do now is make it so that when we come to the app, our session cart ID is created and added as a cookie. This will be the user's cart identifier. Because in our action, we need to check for that session cart id in the cookie in order to get the cart items from the database.
We're adding this logic in a callback in the `auth.ts` file called `authorize`, which invokes when a user needs authorization using middleware (there are directions here - https://authjs.dev/reference/nextjs#authorized ).
Since this callback uses middleware, It will only be called if we add a middleware file and reference the auth file.