Add ability to remove an item from the cart by creating a new action
Edit the add-item-to-cart button to show a quantity and add a remove button.
Now if we click on the + Add to cart button - the item quantity will appear with - and + buttons. If we click on +, the quantity of the item will increase. If we click on -, it will decrease or will be removed from the cart if its count is 1.