Add functionality for persisting session cart
Before/without this change, if you add items to your cart as a guest and then log in, your cart will not persist. To fix this, we're ensuring that the guest cart (session-based) is set as the user cart when the user signs in.
We're doing this by overwriting the user cart on sign in. When a user logs in, we check for their guest sessionCartId. If the guest cart exists, we update the user's cart with the session cart items.

Separate from this case, we also add functionality for handle session updates.